Authentication
Added authentication for all of my currently used endpoints of the api. I have left open the ones who will need an administrative level of access for the time being.

+++ index.js
@@ -1,11 +1,17 @@
 const express = require ('express');
-    bodyParser = require ('body-parser');
+    //bodyParser = require ('body-parser');
     uuid = require ('uuid');
 
 const app = express();
 const mongoose = require('mongoose');
 const Models = require('./models.js');
-app.use(bodyParser.json());
+//app.use(bodyParser.json());
+app.use(express.json());
+app.use(express.urlencoded({extended: true}));
+
+let auth = require('./auth.js')(app);
+const passport = require ('passport');
+require('./passport.js');
 
 const Movies = Models.Movie;
 const Users = Models.User;
@@ -15,7 +21,7 @@ const Heroes = Models.Heroes
 
 mongoose.connect('mongodb://127.0.0.1:27017/MovieDB', {useNewUrlParser: true, useUnifiedTopology: true});
 
-//2.8 getting all users information
+//2.8 getting all users information (need to impliment admin)
  app.get('/users', (req, res) => {
     Users.find()
     .then((users) => {
@@ -27,7 +33,7 @@ mongoose.connect('mongodb://127.0.0.1:27017/MovieDB', {useNewUrlParser: true, us
     });
  });
 
- //2.8 get a user by their name (may not be needed)
+ //2.8 get a user by their ID (need to impliment admin)
  app.get('/users/:userID', (req, res) => {
  Users.findOne({_id: req.params.userID})
  .then ((user) => {
@@ -39,9 +45,9 @@ mongoose.connect('mongodb://127.0.0.1:27017/MovieDB', {useNewUrlParser: true, us
  });
 });
 
- //2.8 *(getting a json of all movies)
- app.get('/movies', (req, res) => {
-    Movies.find()
+ //2.8 ~*(getting a json of all movies)
+ app.get('/movies', passport.authenticate('jwt', {session: false}), async (req, res) => {
+    await Movies.find()
     .then((movie) => {
         res.status(200).json(movie);
     })
@@ -51,8 +57,8 @@ mongoose.connect('mongodb://127.0.0.1:27017/MovieDB', {useNewUrlParser: true, us
     });
  });
 
- //2.8 *(getting a movie via its title)
- app.get('/movies/:movieID', (req, res) => {
+ //2.8 ~*(getting a movie via its ID)
+ app.get('/movies/:movieID', passport.authenticate('jwt', {session: false}), (req, res) => {
     Movies.findOne({_id: req.params.movieID})
     .then ((movie) => {
         if (movie) {
@@ -67,8 +73,8 @@ mongoose.connect('mongodb://127.0.0.1:27017/MovieDB', {useNewUrlParser: true, us
     });
  });
 
- //2.8 *get movie via genre (marvel phase) name.
- app.get('/movies/Genre/:GenreName', (req, res) => {
+ //2.8 ~*get movie via genre (marvel phase) name.
+ app.get('/movies/Genre/:GenreName', passport.authenticate('jwt', {session: false}), (req, res) => {
     Movies.findOne({GenreName: req.params.Name})
     .then ((genre) => {
         if (genre) {
@@ -83,8 +89,8 @@ mongoose.connect('mongodb://127.0.0.1:27017/MovieDB', {useNewUrlParser: true, us
     });
  });
 
- //2.8 *get director information.
- app.get('/movies/director/:DirectorName', (req, res) => {
+ //2.8 ~*get director information.
+ app.get('/movies/director/:DirectorName', passport.authenticate('jwt', {session: false}), (req, res) => {
     Movies.findOne({DirectorName: req.params.Name})
     .then ((director) => {
         if (director) {
@@ -99,8 +105,8 @@ mongoose.connect('mongodb://127.0.0.1:27017/MovieDB', {useNewUrlParser: true, us
     });
  });
 
- //2.8 Get movies that a requested hero is in.
-app.get('/movies/Heroes/:Heroes', (req, res) => {
+ //2.8 ~Get movies that a requested hero is in.
+app.get('/movies/Heroes/:Heroes', passport.authenticate('jwt', {session: false}), (req, res) => {
     Movies.find({Heroes: req.params.Heroes})
     .then((movies) => {
         res.status(200).json(movies)
@@ -111,7 +117,7 @@ app.get('/movies/Heroes/:Heroes', (req, res) => {
     });
 });
 
-//2.8 Updating a movie.
+//2.8 Updating a movie.          (Need to impliment admin)
 app.put('/movies/:movieID', async (req, res) => {
     await Movies.findOneAndUpdate({_id: req.params.movieID},
         { $set:
@@ -135,8 +141,11 @@ app.put('/movies/:movieID', async (req, res) => {
         });
 });
 
- //2.8 *(Updating a Users information)
-app.put('/users/:userID', async (req, res) => {
+ //2.8 ~*(Updating a Users information)
+app.put('/users/:userID', passport.authenticate('jwt', {session: false}), async (req, res) => {
+    if(req.user.id !== req.params.userID){
+        return res.status(400).send('Permission denied');
+    }
     await Users.findOneAndUpdate({_id: req.params.userID}, 
         { $set:
         {
@@ -156,8 +165,11 @@ app.put('/users/:userID', async (req, res) => {
     })    
 });
 
- //2.8 *(Adding a users favorite movies)
-app.post('/users/:Username/movies/:MovieID', async (req, res) => {
+ //2.8 ~*(Adding a users favorite movies)
+app.post('/users/:Username/movies/:MovieID', passport.authenticate('jwt', {session: false}), async (req, res) => {
+    if(req.user.Username !== req.params.Username){
+        return res.status(400).send('Permission denied');
+    }
     await Users.findOneAndUpdate(
         {Username: (req.params.Username)}, 
         {$addToSet: {FavoriteMovies: req.params.MovieID}},
@@ -174,8 +186,8 @@ app.post('/users/:Username/movies/:MovieID', async (req, res) => {
             });
     });
 
-//2.8 *(Adding a user)
-app.post('/users', (req, res) => {
+//2.8 ~*(Adding a user)
+app.post('/users', passport.authenticate('jwt', {session: false}), (req, res) => {
     console.log(req.body)
     Users.findOne({Username: req.body.Username})
     .then ((user) => {
@@ -201,7 +213,7 @@ app.post('/users', (req, res) => {
     });
 });
 
-//2.8 Adding a new Movie to the DB
+//2.8 Adding a new Movie to the DB   (Need to impliment admin)
 app.post('/movies', (req, res) => {
     console.log(req.body)
     Movies.findOne({Title: req.body.Title})
@@ -231,8 +243,11 @@ app.post('/movies', (req, res) => {
        });
     });
 
-//2.8 *(Deleting a user)
-app.delete('/users/:userID', (req, res) => {
+//2.8 ~*(Deleting a user)
+app.delete('/users/:userID', passport.authenticate('jwt', {session: false}), (req, res) => {
+    if(req.user.id !== req.params.userID){
+        return res.status(400).send('Permission denied');
+    }
     Users.findOneAndRemove({_id: req.params.userID})
     .then((user) => {
         if (!user) {
@@ -247,7 +262,7 @@ app.delete('/users/:userID', (req, res) => {
     });
 });
 
-//2.8 Deleting a movie
+//2.8 Deleting a movie                  (Need to impliment Admin)
 app.delete('/movies/:movieID', (req, res) => {
     Movies.findOneAndRemove({_id: req.params.movieID})
     .then((movie) => {
@@ -263,8 +278,11 @@ app.delete('/movies/:movieID', (req, res) => {
     });
 });
 
-//2.8 *Deleting a favorite movie from a users favorite movie array
-app.delete('/users/:Username/movies/:MovieID', (req, res) => {
+//2.8 ~*Deleting a favorite movie from a users favorite movie array
+app.delete('/users/:Username/movies/:MovieID', passport.authenticate('jwt', {session: false}), (req, res) => {
+    if(req.user.Username !== req.params.Username){
+        return res.status(400).send('Permission denied');
+    }
     Users.findOneAndUpdate(
         {Username: req.params.Username}, 
         {$pull: {FavoriteMovies: req.params.MovieID} },
